Prefer using py.exe Python launcher on Windows
On Windows, Python installations come with a launcher called "py.exe"
by default. This launcher locates Python interpreters and allows users
to specify whether to use Python 2 or 3 when running scripts. It's
generally more robust than launching "python.exe" directly.
https://www.python.org/dev/peps/pep-0397/ has more details.

This commit modifies "quadplay.cmd" to check if "py.exe" is present in
the user's path, and uses it to run the quadplay Python script if it
is. This accomplishes two things:
- it allows users with both Python 2 and 3 installations to launch
  quadplay
- it avoids an issue where quadplay.cmd accidentally runs the Windows
  Store python.exe shim and silently exits